Remove all mentions of 'stale relay' flow
There is no real reason to remove the relays from the Hub, as
that is the owner of a stake who is most interested in removing
a relay if not used.

As clients don't read the state but the events over last N blocks,
stale relays won't interfere with a clients' experience.