Add 'RelayMaxNonce' to relay request
It is possible for a relay to delay the relayed transaction indefinitely
by giving it a very high nonce.
The client's responsibility is then to decide (wisely) what is the
maximum relay nonce he will accept.
Higher nonce tx is be considered invalid. The client keeps trying other
relays. That is expected to trigger a race between the discarded high
nonce transaction and another one, with only one of them able to make it
to the blockchain.

On the server side, getting a 'RelayMaxNonce' lower than known
'lastNonce' leads to immediate error.