Add 'solidity-coverage' node command to the project
Current 'solidity-coverage' does not support solidity 0.5.0 and up, so
it relies on patcking the 'parser.js' file in the meantime.

Also, the "instrumented" version of the RelayHub contract is too large,
so we need to add a large gas limit to it's migration sctipt, as well as
run the ganache for coverage with the "allowUnlimitedContractSize" flag.

Note: need to fix "skipFiles" behavior as it does not seem to work.
Note 2: Many tests fail in coverage mode, which must be related to a
change in the number of logs/amounts of gas consumed, etc. Needs fixing.