Add cross compile task in server makefile
Uses karalabe/xgo for cross compiling the relay server for darwin, windows, and linux. Note that the vanilla go cross compiling does not work here, since we have dependencies with native extensions.

As part of this, I had to move two go files from the root of the src directory. By convention, go requires that all code is in a subfolder of the main src folder - and xgo would not detect the files to compile otherwise.