Add destination parameter to rhub withdraw
If a recipient wants to withdraw funds from the rhub, the contract itself needs to call the `withdraw` function on the hub, which transfers funds to the contract. This forces the contract to implement a fallback payable function (ie to accept funds) in order to receive that transfer. Adding a fallback payable function is generally not a good idea, unless there is a mechanism in place to process arbitrary payments made to the contract.

To fix this, we can add a `dest` parameter to rhub's `withdraw`, which indicates who actually receives the withdrawn funds. This removes the requirement for any recipient to have to accept (and manage) funds from any source. Old behaviour can be preserved by just specifying the contract's own address as `dest`.