Add dev mode to relayer server (#199)
* Add dev mode to relayer server

Adds a new bool flag -DevMode (defaults to false). When set, the relayer does not cache its account nonce, and does not try resending unconfirmed txs. This is useful when running unit tests against a relayer instance using ganache: if the test suite rollbacks the blockchain state on each test, the relayer account nonce falls out of sync. This mode prevents this, and allows having a long-running relayer for the whole test suite.

* Check that chain ID is >1000 for allowing devmode

* Remove shortSleep in favour of devMode