RelayClient: should check if "canRelay()" before calling a relay.
The client should check if the transaction it generates is valid
on-chain, before sending it to a relay.

This is done by calling the same `RelayHub.canRelay()` method used by
the relay itself to check the transaction validity.

Without this check, in case the transaction would fail, the relay would
reject it. But in this case, the client would continue to try and send
the transaction through all relays, instead of stopping after first
failure.

When signing a transaction with Metamask, this also has the side-effect
of requiring the user to confirm signing the transaction again for each
such relay access attempt.