Make ECDSA recover EIP-155-avare
The problem is now we just have to migrate to Istanbul because CHAINID
opcode was added there. Storing chainID value in library did not make
sense.
Also, entire ECDSA library had to be checked in and midified. This
could, in theory, be done without accessing the ChainID on-chain, but it
feels like it defeats the purpose.