Add BaseGsnAware baseclass
Paymaster and RelayRecipient are different entities, each knows about a
RelayHub.
But if a single entity has to inherit them both, it must still have a
single relayHub for both purposes.
Without this base class, there are two members named `relayHub`, and its
very confusing as to which one is in use.

So we add a base contract named `BaseGsnAware` which only has the
`relayHub` member for that purpose.