Separete RelayServer into components
Mostly preserving the existing functionality, code fragments and tests
logic, create 2 new components:
- Registration Manager: handles Stake Manger's events, registers if
needed or forced by the server
- Transaction Manager: handles sending and resending transactions,
nonces and transaction storage

Also, in tests, added the 'ServerTestUtils' to extract non-test
functions from the '.test.ts' files. Note that weird-looking 'interface'
declarations in tests are not added, but replace 'any' objects.

Minor in-the-go changes:
1. Fix address2topic when address has uppercase characters
2. Remove 'topicsForManagers' to be a utility function
3. Wider the usage of Contract Interactor in the server. Note that this
means it is now a shared component and should not reference client code.
4. Use 'loglevel' library instead of 'debug' flag ins server. Should be
done in the client as well.