OG-656: Consider a transaction 'pending' for N blocks after mining
There is an inherent race condition between server sending a transaction
and observing the result in emitted event logs.

Up until now, the code considered the transaction to stop being
'pending' the moment it is mined. However, it appears to take more time
to get the RPC node to provide a satisfying response to
'getTransactionReceipt' and 'getPastEvents', causing the server to send
multiple identical transactions.

Make 'recentActionAvoidRepeatDistanceBlocks' a parameter