OG-1005: WIP: Allow adjusting the RelayRequest for the winning PingResponse
The user experience of rejecting a Relay Server because its gas price
expectations did not match ours is frustrating.

Instead, we can allow the client-side code to agree to the Relay's terms
unconditionally and that should decrease the number of such rejections
significantly.

The risk of doing so is pretty low as long as we keep any kind of check.